new homepage copy
Breaking this down like Stirpe's new homepage, http://stripe.com, we can
create 4 vertically-expanded sections with the copy in this PR.

For the second section, the one with a 5x2 grid of questions (in desktop
mode), I was envisioning that when a user presses a question there would
be text beneath that question explaining to him or her how we can help.

Stripe is a great exemplar website because they have great fonts and
colors, and use of icons throughout the page. Their page also has a lot
of breathing room - which is something you have commented on.